+++ src/xeda/flows/dc/__init__.py
@@ -213,11 +213,7 @@ class Dc(AsicSynthFlow):
         @field_validator("platform", mode="before")
         @classmethod
         def _validate_platform(cls, value):
-            if isinstance(value, str) and not value.endswith(".toml"):
-                return AsicsPlatform.from_resource(value)
-            elif isinstance(value, (str, Path)):
-                return AsicsPlatform.from_toml(value)
-            return value
+            return AsicsPlatform.from_setting(value)
 
     def init(self):
         assert isinstance(self.settings, self.Settings)